// src/app/shared/models/factory.model.ts
import { City } from './city.model';

export class Factory {
    idFactory?: number;
    name?: string;
    address?: string;
    neighborhood?: string;
    cep?: string;
    state?: string;
    cnpj?: string;
    latitude?: string;
    longitude?: string;
    perimeter?: string;
    active?: boolean;
    city?: City;
}